chore(host): move host from legacy to pages
