--- src/legacy/routes/Host/index.ts
@@ -1,1 +1,0 @@
-export { default } from './Host';
